refactoring runAll for China
