+++ src/cli/commands/index.js
@@ -5,7 +5,6 @@ const info = require('./info')
 const create = require('./create')
 const dev = require('./dev')
 const registry = require('./registry')
-const template = require('./template')
 
 module.exports = {
   login,
@@ -14,6 +13,5 @@ module.exports = {
   info,
   create,
   dev,
-  registry,
-  template
+  registry
 }
